Integrate SP3 lib to ecosystem
One solution to having a single cargo workspace for this ecosystem is either

* have SP3 into this crate: which facilitates future developments
* move all applications (non libraries, ie., library users like rinex-cli)
into an external workspace

Having SP3 into this is probably the best solution.
For example "rinex-qc" makes no sense without SP3 and RINEX knowledge.
"rinex-cli" is a heavy user of "rinex-qc" (one of the two major opmodes).
I think that answers the question.

Signed-off-by: Guillaume W. Bres <guillaume.bressaix@gmail.com>