Verify that epochs have a minimum number of fields
When parsing an epoch we want to ensure that we can at least read the
date and time fields. This function will end up consuming non-epoch
lines occasionally and we want to be sure that we bail early with an
error rather than either barreling ahead and triggering an "invalid
Gregorian date" panic or (even worse) assuming the bad parse is
valid.

Signed-off-by: Jason Gerecke <killertofu@gmail.com>