Handle (ignore) observation epochs containing event flags (#226)
* Fix meteorological is_new_epoch version check

* Verify that epochs have a minimum number of fields

When parsing an epoch we want to ensure that we can at least read the
date and time fields. This function will end up consuming non-epoch
lines occasionally and we want to be sure that we bail early with an
error rather than either barreling ahead and triggering an "invalid
Gregorian date" panic or (even worse) assuming the bad parse is
valid.

* epoch: Add formatting checks to cargo tests

Use the "format" function instead of the "format!" macro to verify that the
module produces correctly-formatted epoch strings.

* Make EpochFlag processing the responsibility of the observation module

Epoch flags are only used by observations, so it does not make sense to
have their processing done within the epoch module. Moves processing and
testing of the parsing and formatting of the flags over to the
observation module.

* Split parsing of normal observations and events into dedicated functions

We don't yet actually *do* anything with events, but we can at least
return an Error to prevent the rest of the system from trying to
treat them as normal observations.

* Handle observations without significant epoch

Event flags 2 - 5 allow the epoch fields to be left blank if they are
not "significant".

* Move Epoch Flag to Observation RINEX dedicated section
* add one comment

---------

Signed-off-by: Jason Gerecke <killertofu@gmail.com>
Signed-off-by: Guillaume W. Bres <guillaume.bressaix@gmail.com>
Co-authored-by: Guillaume W. Bres <guillaume.bressaix@gmail.com>